Edit some seeders to only delete data if they exist.

+++ backend/db/seeders/20240702034832-member-added.js
@@ -64,8 +64,10 @@ module.exports = {
       where: { email: "john.smith@gmail.com" }
     });
 
-    await queryInterface.bulkDelete(options, {
-      userId: organizer.id
-    });
+    if (organizer) {
+      await queryInterface.bulkDelete(options, {
+        userId: organizer.id
+      });
+    }
   }
 };
