+++ backend/db/seeders/20240702035418-group-image-added.js
@@ -55,8 +55,10 @@ module.exports = {
       where: { name: "Evening Tennis on the Water" }
     });
 
-    await queryInterface.bulkDelete(options, {
-      groupId: group1.id
-    });
+    if (group1) {
+      await queryInterface.bulkDelete(options, {
+        groupId: group1.id
+      });
+    }
   }
 };
